Create MarvelCrypto class to encapsulate md5 cryptography
